% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\docType{data}
\name{RSet}
\alias{RSet}
\title{The set reference class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RSet
}
\description{
The set reference class gives
}
\author{
Yukai Yang, \email{yukai.yang@statistik.uu.se}
}
\keyword{class}
