% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6DS.R
\docType{package}
\name{R6DS}
\alias{R6DS}
\alias{R6DS-package}
\title{R6DS: a package implementing some useful data stuctures by using the R6 reference class.}
\description{
Write something here
}
\details{
And here.
}
\section{Author and Maintainer}{

Yukai Yang

Department of Statistics, Uppsala University

\email{yukai.yang@statistik.uu.se}
}

\section{Linked List}{

\code{\link{SimModel1}} simulate from the type one state-space model on the Stiefel manifold.
}

\section{Stack}{

\code{\link{FilterModel1}} filtering algorithm for the type one model.
}

\section{Queue}{

\code{\link{runif_sm}} sample from the uniform distribution on the Stiefel manifold.
}

\section{Set}{

\code{\link{version}} shows the version number and some information of the package.
}

\section{Dictionary}{

\code{\link{version}} shows the version number and some information of the package.
}

\section{Binary Search Tree (BST)}{

\code{\link{version}} shows the version number and some information of the package.
}

